migrations issues fixed

+++ database/migrations/2022_05_13_091200_create_admins_email_verification_table.php
@@ -15,7 +15,7 @@ class CreateAdminsEmailVerificationTable extends Migration
     {
         Schema::create('admins_email_verifications', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('admin_id')->unsigned();
+            $table->bigInteger('admin_id')->unsigned();
             $table->string('token');
             $table->boolean('is_verified')->default(false);
